Starnavi/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
better-profanity==0.7.0
cffi==1.17.1
cryptography==43.0.1
Django==5.1.2
django-environ==0.11.2
django-ninja==1.3.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
idna==3.10
iniconfig==2.0.0
joblib==1.4.2
numpy==2.1.2
packaging==24.1
pluggy==1.5.0
profanity-check==1.0.3
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-jose==3.3.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
